+++ src/components/Users/User/User.module.scss
@@ -19,8 +19,8 @@
     border-radius: $space-l;
 
     text-align: center;
-
     color: $color-white;
+
     background-color: $color-blue;
 
     &:hover {
