+++ src/pages/Login/Login.tsx
@@ -22,14 +22,12 @@ const Login: React.FC = () => {
     axios
       .post(`/authenticate`, loginData)
       .then((response) => {
-        console.log("aaaa");
         if (response.status === 200) {
           localStorage.setItem('token', response.data.accessToken);
           navigate('/profile');
         }
       })
       .catch((err) => {
-        console.log(err.status);
         setInputError(true);
       });
   };
